lint

+++ src/link/soem/soem_handler.hpp
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 03/12/2022
+// Last Modified: 06/12/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -166,7 +166,7 @@ class SOEMHandler final {
     if (ec_slave[0].state != EC_STATE_SAFE_OP) {
       spdlog::error("One ore more slaves did not reach safe operational state: {}", ec_slave[0].state);
       ec_readstate();
-      for (size_t slave = 1; slave <= ec_slavecount; slave++)
+      for (size_t slave = 1; slave <= static_cast<size_t>(ec_slavecount); slave++)
         if (ec_slave[slave].state != EC_STATE_SAFE_OP)
           spdlog::error("Slave[{}]: {} (State={:#02x} StatusCode={:#04x})", slave, ec_ALstatuscode2string(ec_slave[slave].ALstatuscode),
                         ec_slave[slave].state, ec_slave[slave].ALstatuscode);
@@ -195,7 +195,7 @@ class SOEMHandler final {
       if (remaining == 0) {
         spdlog::error("One ore more slaves are not responding: {}", ec_slave[0].state);
         ec_readstate();
-        for (size_t slave = 1; slave <= ec_slavecount; slave++)
+        for (size_t slave = 1; slave <= static_cast<size_t>(ec_slavecount); slave++)
           if (ec_slave[slave].state != EC_STATE_SAFE_OP)
             spdlog::error("Slave {} State={:#02x} StatusCode={:#04x} : {}", slave, ec_slave[slave].state, ec_slave[slave].ALstatuscode,
                           ec_ALstatuscode2string(ec_slave[slave].ALstatuscode));
